backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.0
python-dotenv==1.0.0
motor==3.1.1
pydantic==1.10.12